// steins-feed-ui-vite/src/client/types.gen.ts
export type Feed = {
    id: number;
    title: string;
    link: string;
    language: Language | null;
    tags: Array<Tag>;
    displayed: boolean;
};

export type HttpValidationError = {
    detail?: Array<ValidationError>;
};

export type Item = {
    id: number;
    title: string;
    link: string;
    published: string;
    summary: string | null;
    feed: Feed;
    like: LikeStatus | null;
    magic: number | null;
    surprise: number | null;
};

export type Language = 'English' | 'French' | 'German' | 'Swedish';

export type LikeStatus = 1 | 0 | -1;

export type Tag = {
    id: number;
    name: string;
};